requirements.txt
google-cloud
pyyaml